Version 1.0.1 - See CHANGELOG.md for more detail.
