+++ CHANGELOG.md
@@ -1,0 +1,14 @@
+# ChalkFlags Changelog
+
+### Version 1.0.0
+Initial version of ChalkFlags.
+
+- Supports one flag per word
+
+---
+
+### Version 1.0.1
+Rewrite of everything.
+
+- Multiple flags per word
+- Revision of the whole engine
