+++ demo/helloworld.js
@@ -1,0 +1,5 @@
+import cf from '../index.js';
+
+const e = new cf();
+
+console.log(e.parse('r/bd/Hello u/i/World'));
